% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{calAvgEmissionMat}
\alias{calAvgEmissionMat}
\title{Use the list of EmissionCounts for different chromosomes
to estimate the emission probabilisties from the learned data.}
\usage{
calAvgEmissionMat(
  listOfEmissionCount,
  mods = c("ATAC", "H3K27ac", "H3K27me3", "H3K4me1", "H3K9me3"),
  states = ordStates
)
}
\arguments{
\item{listOfEmissionCount}{list of emission counts from the function
of getEmissionCount.Chrom}

\item{mods}{vector of modularities.}

\item{states}{vector of states}
}
\value{
data.frame with columns of
raw counts per mod, totalcount for states, and the avgMat
per mod.
}
\description{
Use the list of EmissionCounts for different chromosomes
to estimate the emission probabilisties from the learned data.
}
