% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capMat}
\alias{capMat}
\title{capMat}
\usage{
capMat(mat, lowQ = 0.005, upQ = 0.995)
}
\arguments{
\item{mat}{A matrix}

\item{lowQ}{Lower quantile}

\item{upQ}{Upper quantile}
}
\description{
Cap the values of a matrix
}
