% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{combineHeatmap}
\alias{combineHeatmap}
\title{Combine two ComplexHeatmaps.}
\usage{
combineHeatmap(chm1, chm2, legend_direct = "bottom", legend_merge = TRUE)
}
\arguments{
\item{chm1}{ComplextHeatmap 1}

\item{chm2}{ComplextHeatmap 2}

\item{legend_direct}{string, "bottom" by default}

\item{legend_merge}{boolean, TRUE by default}
}
\value{
ComplextHeatmap
}
\description{
Combine two ComplexHeatmaps.
}
