% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{downsampleSeurat}
\alias{downsampleSeurat}
\title{depend on get.downsample.fun in the same file}
\usage{
downsampleSeurat(seu, groupBy = "cluster_id", minNum = 1000, maxNum = 2000)
}
\description{
depend on get.downsample.fun in the same file
}
