% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{extendGR}
\alias{extendGR}
\title{Extend upstream and downsteam of Genomic ranges}
\usage{
extendGR(g, up = 500, down = 500)
}
\arguments{
\item{g}{genomic ranges}

\item{up}{upstream extSize, positive integer}

\item{down}{downstream extSize, positive integer}
}
\value{
a new Genomic ranges with extSizes.
}
\description{
Extend upstream and downsteam of Genomic ranges
}
