% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastCorOnRow.chunk}
\alias{fastCorOnRow.chunk}
\title{fastCorOnRow.chunk}
\usage{
fastCorOnRow.chunk(
  p1byc,
  p2byc,
  p12pairs,
  chunkSize = 1e+06,
  corMethod = c("spearman", "pearson")
)
}
\arguments{
\item{p1byc}{scaled mat with row and col names
ncol(p1byc) = ncol(p2byc)}

\item{p2byc}{scaled mat with row and col names
ncol(p1byc) = ncol(p2byc)}
}
\description{
fastCorOnRow.chunk
}
