% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastPearCorOnRow.scaled}
\alias{fastPearCorOnRow.scaled}
\title{fastPearCorOnRow.scaled}
\usage{
fastPearCorOnRow.scaled(mat1, mat2)
}
\arguments{
\item{mat1, }{row scaled mat, same shape as mat2}

\item{mat2, }{row scaled mat, same shape as mat1}
}
\description{
fastPearCorOnRow.scaled
}
