% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastSpearCorOnRow.distinctRank}
\alias{fastSpearCorOnRow.distinctRank}
\title{fastSpearCorOnRow.distinctRank}
\usage{
fastSpearCorOnRow.distinctRank(mat1, mat2, tiesMethod = "average")
}
\arguments{
\item{mat1}{same shape as mat2}

\item{mat2}{same shape as mat1}
}
\description{
fastSpearCorOnRow.distinctRank
}
\examples{

n <- 20000
row.index <- sample(min(nrow(s.rna.mat),nrow(s.atac.mat)), n)
## pearson 
r1 <- fastCorrOnRow(mat1 = s.rna.mat[row.index, ], mat2 = s.atac.mat[row.index, ])
r2 <- vapply(row.index, function(i) {
  cor(x = s.rna.mat[i,], y = s.atac.mat[i,], method = "pearson")
}, FUN.VALUE = 0.1)
plot(x = r1, y = r2, xlab = "factCor", ylab = "cor")
abline(a=0, b=1)
qqplot(x = r1, y = r2)
abline(a=0, b=1)
## spearman
r1 <- fastCorrOnRow(
  mat1 = normalizeUseRankByRow(s.rna.mat[row.index, ]),
  mat2 = normalizeUseRankByRow(s.atac.mat[row.index, ]))
r2 <- vapply(row.index, function(i) {
  cor(x = s.rna.mat[i,], y = s.atac.mat[i,], method = "spearman")
}, FUN.VALUE = 0.1)
plot(x = r1, y = r2, xlab = "factCor", ylab = "cor")
abline(a=0, b=1)
qqplot(x = r1, y = r2)
abline(a=0, b=1)

}
