% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{findOverlaps2}
\alias{findOverlaps2}
\title{Modification of GenomicRanges::findOverlaps by adding
overlap region, width.}
\usage{
findOverlaps2(queryGR, subjectGR)
}
\description{
Modification of GenomicRanges::findOverlaps by adding
overlap region, width.
}
