% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{FindOvlpRegionInB}
\alias{FindOvlpRegionInB}
\title{Find the genomic regions in A that overlap with B}
\usage{
FindOvlpRegionInB(A, B)
}
\description{
Find the genomic regions in A that overlap with B
}
