% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{getAllRcolors}
\alias{getAllRcolors}
\title{Use R colors firstly: colors()
https://r-graph-gallery.com/ggplot2-color.html}
\usage{
getAllRcolors(file = "../meta/R.color.pdf")
}
\description{
Use R colors firstly: colors()
https://r-graph-gallery.com/ggplot2-color.html
}
