% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{getBinarizeMonocle3Obj}
\alias{getBinarizeMonocle3Obj}
\title{Generate Monocle3 object for cicero analysis}
\usage{
getBinarizeMonocle3Obj(
  mat,
  cellinfo = NULL,
  peakinfo = NULL,
  isSparse = T,
  isRowBarcode = T,
  splitby = ":|-"
)
}
\arguments{
\item{mat}{matrix or sparse matrix with rownames (features) and
colnames (cells)}

\item{cellinfo}{data.frame rownames is the same as the colnames of
mat If NULL, will transform colnames of mat to data.frame.}

\item{peakinfo}{data.frame rownames is the same as the rownames of
mat If NULL, will transform rownames of mat to data.frame.}

\item{isSparse}{bool for mat}

\item{isRowBarcode}{bool for mat}

\item{splitby}{characters for spliting peak string}
}
\value{
monocle3::cell_data_set object
}
\description{
In Monocle3, the matrix can be sparse or not.
But the row of the stored matrix will be genes or features
Also the matrix will be saved as dgCMatrix if sparse.
If dgRMatrix is used, will be transformed into dgCMatrix, which
needs to load Matrix package explicitly.
Matrix will be binarzied if they have reads orcounts or not,
and full zero cells and features will be filtered out.
}
\details{
FIXME:
Warning messages:
The x argument of as_tibble.matrix()
must have unique column names if .name_repair
is omitted as of tibble 2.0.0. Using compatibility .name_repair.
}
