% Generated by roxygen2: do not edit by hand
% Please edit documentation in bioR/tool_bigwig.R
\name{getbwSignal}
\alias{getbwSignal}
\title{Get bigwig signals for a given genomic ranges}
\usage{
getbwSignal(bwGR, peakGR, fnbw = identity, fnGroup = mean)
}
\arguments{
\item{bwGR}{use rtracklayer::import to load bigwig file directly.
The results will be genomic ranges object.}

\item{peakGR}{genomic ranges interested
We assumme a name field for the ranges, which will be the names
for the output vector.}

\item{fnbw}{how to handle the overlapped bigwig signal itself
fn: Double => Double, default identity}

\item{fnGroup}{how to handle all the overlapped bigwig signals for
a given region. fn: Vector\link{Double} => Double, default mean}
}
\value{
named numeric vector
The order the vector follows the input of peakGR
}
\description{
Get bigwig signals for a given genomic ranges
}
