% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getcov.mat}
\alias{getcov.mat}
\title{Get coefficicent of variation for each features.
mat: cell by features}
\usage{
getcov.mat(mat)
}
\description{
Get coefficicent of variation for each features.
mat: cell by features
}
