% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{getCPMOfHistones}
\alias{getCPMOfHistones}
\title{Perform CPM for Histone raw feature counts.}
\usage{
getCPMOfHistones(cnt_pbyc, scale.factor = 1e+06)
}
\arguments{
\item{cnt_pbyc}{matrix, feature by group}

\item{scale.factor}{default 1e6}
}
\value{
CPM, feature by group
}
\description{
Perform CPM for Histone raw feature counts.
}
