% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{getCPMOfptRNA}
\alias{getCPMOfptRNA}
\title{Get CPM normalized Paired-Tag gene expressions.}
\usage{
getCPMOfptRNA(fnm, scaleFactor = 1e+06)
}
\arguments{
\item{fnm}{\link{ptPseudoRNAcountfnm}
R's rds file, subclass by gene expression.}

\item{scaleFactor}{used to scaling, default is 1e6 (million)}
}
\value{
matrix, group by feature
}
\description{
BUG:
\itemize{
\item The original data is data.frame, which has to been converted into
matrix. Otherwise, will have weired behaviour on CPM normalization.
(i.e., matrix / vector and data.frame / vector are different?)
\item "|>" piple operator leads to error during noramlization.
\itemize{
\item Error in (function(x) (x * 10^6)) ((function(x) x\link{atac2ptsc$pt, })
(as.matrix(readRDS(ptPseudoRNAcountfnm)))) \%*\%
:non-conformable arguments
}
}
}
