% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{getEmissionCount.Chrom}
\alias{getEmissionCount.Chrom}
\title{Get the emission counts for one chromosome.}
\usage{
getEmissionCount.Chrom(bin2PromoterDistal, bin2State, bin2BinarizeData)
}
\arguments{
\item{bin2PromoterDistal}{data.frame from the function
mapBin2PromoterDistal}

\item{bin2State}{data.frame from the function mapBin2State.Chrom}

\item{bin2BinarizeData}{data.frame from the function
loadChromHMMBinarizedData and focus on one chromsom.}
}
\value{
data.frame with id, state, relaTSS, pdstate,
different modularites and gsize.
Add all the counts for those modularities.
gsize is the total counts for each of the pdstate.
}
\description{
Get the emission counts for one chromosome.
}
