% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{getFullAnnX}
\alias{getFullAnnX}
\title{Get the full sparse matrix from AnnData}
\usage{
getFullAnnX(ann)
}
\value{
sparse matrix (dgR / dgC depends on AnnData)
with barcodes as rownames and features as colnames
}
\description{
Get the full sparse matrix from AnnData
}
