% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{getGeneGR}
\alias{getGeneGR}
\title{From GTF (loaded using rtracklayer::import from gtf / gff\link{3})
GenomicRanges, get the gene GenomeRanges based on its type}
\usage{
getGeneGR(gtf, mcolnms = c("gene_id", "type", "gene_type", "gene_name"))
}
\arguments{
\item{gtf}{GenomicRanges}

\item{mcolnm}{vector of characters metadata for result}
}
\value{
GenomicRanges
}
\description{
From GTF (loaded using rtracklayer::import from gtf / gff\link{3})
GenomicRanges, get the gene GenomeRanges based on its type
}
