% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{getHomerPromoterGR}
\alias{getHomerPromoterGR}
\title{Get Promoter from Homer TSS}
\usage{
getHomerPromoterGR(
  homertssf,
  bufferSize = 2000,
  upstream = 1000,
  downstream = 500
)
}
\arguments{
\item{homertssf}{full path of homer TSS file}

\item{bufferSize}{2000 See details in \link{readHomerTSS}}

\item{upstream}{1000}

\item{downstream}{500}
}
\value{
GenomicRanges object
}
\description{
Get Promoter from Homer TSS
}
