% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{getMetaCol.Seurat}
\alias{getMetaCol.Seurat}
\title{getMetaCol.Seurat}
\usage{
getMetaCol.Seurat(seu, colnm)
}
\value{
array with name
}
\description{
use repeat colnm to get array
instead of data.frame with one column
}
