% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{getMotifMat}
\alias{getMotifMat}
\title{Summarize motif results across multiple groups
into a unified matrix.}
\usage{
getMotifMat(motif_list, field = "p", tf = identity)
}
\arguments{
\item{motif_list}{list of homer results for different groups}

\item{field}{which column for the final matrix}

\item{tf}{function to transform the column
for example, we can use log10 or -log10 for negative
log10 of pvalue.}
}
\value{
matrix groups as rowname, motifs as colnames
}
\description{
We assume that, for each motif result, it's a data.frame
following the formats defined in loadHomer2KnownResults function.
}
