% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvar.R
\name{getNCellOfModality}
\alias{getNCellOfModality}
\title{Get cell number record from a typical modality string like
"H3K4me1-male125:92:33_female105:79:26"}
\usage{
getNCellOfModality(x)
}
\arguments{
\item{x}{modality string}
}
\value{
integer of the total number of cells
}
\description{
Get cell number record from a typical modality string like
"H3K4me1-male125:92:33_female105:79:26"
}
