% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getOrderOfElementsByGroupOrder}
\alias{getOrderOfElementsByGroupOrder}
\title{Get order of elements by group order}
\usage{
getOrderOfElementsByGroupOrder(g, group)
}
\arguments{
\item{g}{vector of group info}

\item{group}{ordered unique names of g}
}
\value{
vector order of g
}
\description{
Get order of elements by group order
}
