% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getOrderOfRowsByOrdersOfCols}
\alias{getOrderOfRowsByOrdersOfCols}
\title{Get order of rows by orders of columns}
\usage{
getOrderOfRowsByOrdersOfCols(enrich, topk = 1)
}
\arguments{
\item{enrich}{data.frame}

\item{topk}{topk to sort the rows}
}
\value{
vector order of rows
}
\description{
Get order of rows by orders of columns
}
