% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvar.R
\name{getPairedTagSubclass2Nmod}
\alias{getPairedTagSubclass2Nmod}
\title{Get PairedTag subclasses' number of cells for different
moadlities.}
\usage{
getPairedTagSubclass2Nmod(ptscMeta)
}
\arguments{
\item{ptscMeta}{see details in \link{loadptscMeta}}
}
\value{
data.frame with columns of
PairedTagSubclassName, different modalities and overall DNA numbers
}
\description{
Use function of \link{getNCellOfModality} to extract the
number information.
}
