% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{getPartAnnX}
\alias{getPartAnnX}
\title{Get part of sparse matrix from AnnData}
\usage{
getPartAnnX(ann, groupby, group)
}
\value{
sparse matrix (dgR / dgC depends on AnnData)
with barcodes as rownames and features as colnames
}
\description{
Get part of sparse matrix from AnnData
}
