% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{groupColsByHierarchical}
\alias{groupColsByHierarchical}
\title{Group columns by hierarchical clustering}
\usage{
groupColsByHierarchical(
  mat,
  dmethod = "pearson",
  hmethod = "ward.D2",
  k = 50,
  sumfn = mean
)
}
\arguments{
\item{mat}{A matrix to be clustered
matrix should be in the form of samples x features}

\item{dmethod}{A method to calculate distance}

\item{hmethod}{A method to calculate hierarchical clustering}

\item{k}{A number of clusters}

\item{sumfn}{A function to summarize the grouped columns
in row-wise (will transform the mat during innter calculation).
Default is rowMeans}
}
\value{
list of enriched matrix, group (vector), and hclust object
}
\description{
Group columns by hierarchical clustering
}
