% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{load_sc2scil}
\alias{load_sc2scil}
\title{Map subclass in snATAC name habit to pt subclass id label habit}
\usage{
load_sc2scil()
}
\value{
data.frame with sc_atac and sc as two columns
}
\description{
Map subclass in snATAC name habit to pt subclass id label habit
}
