% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_atac.R
\name{loadAllCREAsGR}
\alias{loadAllCREAsGR}
\title{Load All one million CREs from ATAC-seq data as Genomic Ranges}
\usage{
loadAllCREAsGR()
}
\description{
Load All one million CREs from ATAC-seq data as Genomic Ranges
}
