% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvar.R
\name{loadAllenTFOrd}
\alias{loadAllenTFOrd}
\title{Load 260 TFs in order used by Allen for their paper Fig.5.d
The raw is directly downloaded from the paper supplemantary file.
Then manually remove header and convert to txt file.}
\usage{
loadAllenTFOrd()
}
\value{
array of TF gene symbols
}
\description{
Load 260 TFs in order used by Allen for their paper Fig.5.d
The raw is directly downloaded from the paper supplemantary file.
Then manually remove header and convert to txt file.
}
