% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{loadAllppec}
\alias{loadAllppec}
\title{Read all the subclasses' positive proximal-enhancer connections}
\usage{
loadAllppec(ppecd)
}
\arguments{
\item{ppecd}{\link{snATACppecd}}
}
\value{
list of data.frame
}
\description{
Read all the subclasses' positive proximal-enhancer connections
}
