% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_atac.R
\name{loadATACPseudoBulkMat}
\alias{loadATACPseudoBulkMat}
\title{Load ATAC pseudo-bulk CPM / Count as matrix.}
\usage{
loadATACPseudoBulkMat(
  f,
  pd = NULL,
  env = "sa2",
  conda = "/home/szu/mambaforge/bin/conda"
)
}
\arguments{
\item{f}{\link{snATACPMpbyc} or \link{snATACountpbyc}}

\item{pd}{pandas module from reticulate::import
if NULL, will set up it here.}

\item{env}{conda env used for pd, "sa2" by default}

\item{conda}{conda bin path}
}
\value{
matrix with both rownames and colnames.
}
\description{
Load ATAC pseudo-bulk CPM / Count as matrix.
}
