% Generated by roxygen2: do not edit by hand
% Please edit documentation in bioR/tool_bigwig.R
\name{loadBigWigGR}
\alias{loadBigWigGR}
\title{Load Bigwig file into a Genomic range object.}
\usage{
loadBigWigGR(bwfnm)
}
\arguments{
\item{bwfnm}{bigwig file name}
}
\description{
Load Bigwig file into a Genomic range object.
}
