% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{loadChromHMMBinarizedData}
\alias{loadChromHMMBinarizedData}
\title{Read ChromHMM Binarized Data}
\usage{
loadChromHMMBinarizedData(prefix, chrs)
}
\arguments{
\item{prefix}{path prefix
The binarized file is typically named as
prefix_\link{chr}_binary.txt}

\item{chrs}{vector of chromosomes for to read}
}
\value{
list of data.frames with columns
id (start from 1) followed by different modules.
The names of the data.frames are chrs. If some chrs have no
binary files, will skip the corresponding chrs.
If all the chrs have no binary files, return NULL.
}
\description{
Read ChromHMM Binarized Data
}
