% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{loadCRE2Gene}
\alias{loadCRE2Gene}
\title{Get distal2gene (by most correlations) from correlation of
pseudobulk-level RNA-seq and ATAC-seq.}
\usage{
loadCRE2Gene(fnm, genes = NULL, distals = NULL)
}
\arguments{
\item{fnm}{\link{ppdcWithCorfnm}}

\item{genes}{vector of genes to keep, can be NULL}

\item{distals}{vector of CREs to keep, can be NULL}
}
\value{
data.frame
two columns: distal and gene
}
\description{
Get distal2gene (by most correlations) from correlation of
pseudobulk-level RNA-seq and ATAC-seq.
}
