% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{loadGRfromBed}
\alias{loadGRfromBed}
\title{Transform Bed to GenomicRanges
Bed is from a bedFile or a data.frame}
\usage{
loadGRfromBed(
  bedFile = NULL,
  beds = NULL,
  header = FALSE,
  sep = "\\t",
  colnms = c("chr", "start", "end", "name")
)
}
\arguments{
\item{bedFile}{str or NULL}

\item{beds}{data.frame or NULL
for data.frame, at least 3 columns
(first 3 columns are chrom, start end)}

\item{header}{boolean, default FALSE}

\item{sep}{str, default tab}

\item{colnms}{List\link{str}, default is
c("chr", "start", "end", "name")}
}
\value{
GenomicRanges
}
\description{
Transform Bed to GenomicRanges
Bed is from a bedFile or a data.frame
}
