% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_bigwig.R
\name{loadHistoneBigWig}
\alias{loadHistoneBigWig}
\title{Load one histone's bigwig as Genomic ranges}
\usage{
loadHistoneBigWig(sc, h)
}
\arguments{
\item{sc}{PairedTag subclass name}
}
\description{
Load one histone's bigwig as Genomic ranges
}
