% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{loadHomer2KnownResults}
\alias{loadHomer2KnownResults}
\title{Load the known motifs learned from Homer2.
The file end with knownResults.txt will be handled here.}
\usage{
loadHomer2KnownResults(resultd, rmdupmotif)
}
\arguments{
\item{resultd}{where the homer2 results are saved}

\item{rmdupmotif}{default TRUE}
}
\value{
data.frame
}
\description{
Load the known motifs learned from Homer2.
The file end with knownResults.txt will be handled here.
}
