% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\name{loadppdc}
\alias{loadppdc}
\title{Read positive proximal-distal connections of bedpe format.}
\usage{
loadppdc(fnm)
}
\arguments{
\item{fnm}{\link{ciceroPosProximalDistalConnfnm}}
}
\value{
data.frame with columns defined
}
\description{
Read positive proximal-distal connections of bedpe format.
}
