% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvar.R
\name{loadPromoter}
\alias{loadPromoter}
\title{Load default promoter as GenomicRanges object.}
\usage{
loadPromoter(fnm = promoterfnm)
}
\value{
GenomicRanges object
}
\description{
Load default promoter as GenomicRanges object.
}
