% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_atac.R
\name{loadsrtCREnameInPairedTag}
\alias{loadsrtCREnameInPairedTag}
\title{Get ATAC Peaks' located in PairedTag region}
\usage{
loadsrtCREnameInPairedTag()
}
\value{
vector of string
}
\description{
Get ATAC Peaks' located in PairedTag region
}
