% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{map2snATACscstr}
\alias{map2snATACscstr}
\title{map subclass name to snATAC subclass format}
\usage{
map2snATACscstr(sc)
}
\arguments{
\item{sc}{vector of string}
}
\value{
vector of string
}
\description{
map subclass name to snATAC subclass format
}
