% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{mapBin2PromoterDistal.Chrom}
\alias{mapBin2PromoterDistal.Chrom}
\title{Map Genomic bins to promoer or distals based on homer TSS.}
\usage{
mapBin2PromoterDistal.Chrom(
  chrom2Bin,
  homerTSS,
  upSize = 1000,
  downSize = 1000,
  ignoreStrand = T
)
}
\arguments{
\item{chrom2Bin}{data.frame got from paritionChrom
Here chrom2Bin is for one chromosome.}

\item{homerTSS}{data.frame got from readHomerTSS}

\item{upSize}{integer default 1000 up to TSS}

\item{downSize}{integer default 1000 down to TSS}

\item{ignoreStrand}{bool default TRUE
chrom2Bin has no strand information.}
}
\value{
data.frame with the columns of
id, chrom2Bin's id; relaTSS, "D" for distals, "P" for
promoters.
The rownames are "id1", "id2"...
}
\description{
Map Genomic bins to promoer or distals based on homer TSS.
}
