% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{mapBin2State.Chrom}
\alias{mapBin2State.Chrom}
\title{Map genomic bins to chromHMM states.}
\usage{
mapBin2State.Chrom(segment, chrom2Bin, ignoreStrand = T)
}
\arguments{
\item{segment}{data.frame genomic states read from chromHMM}

\item{chrom2Bin}{data.frame from paritionChrom for one chromsome}

\item{ignoreStrand}{bool default TRUE sine chrom2Bin has no
strand information.}
}
\value{
data.frame with the columns id (chromBin id) and state
defined in chromHMM. The rownames are "id1", "id2", ...
}
\description{
Map genomic bins to chromHMM states.
}
