% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{mapfc2fcsc}
\alias{mapfc2fcsc}
\title{Map Histone raw feature counts to subclass-level
Non-neuronal groups are supertypes, so we need a map.}
\usage{
mapfc2fcsc(fc, sp2scMap)
}
\description{
Map Histone raw feature counts to subclass-level
Non-neuronal groups are supertypes, so we need a map.
}
