% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{mapGmatToGeneName}
\alias{mapGmatToGeneName}
\title{mapGmatToGeneName}
\usage{
mapGmatToGeneName(gmat, gencode, cells, fn = base::rowMeans)
}
\value{
data.table, cell by gene
}
\description{
mapGmatToGeneName
}
