% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{partitionChrom}
\alias{partitionChrom}
\title{Given one chrom size, partition it into bins.}
\usage{
partitionChrom(chrName, chromSize, binSize = 200, excludeLastPartialBin = T)
}
\arguments{
\item{chrName}{one chromosome name}

\item{chromSize}{integer size of the chrom}

\item{binSize}{integer}

\item{excludeLastPartialBin}{default TRUE
ChromHMM will exclude the last bin if it's small than binSize}
}
\value{
data.frame of chr, start, end, bin id (from 1)
chrom labels following bed format, i.e.,
start from 0, and left closed and right open
for example [0, 1), 1st position.
}
\description{
Given one chrom size, partition it into bins.
}
