% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{prepareDotPlot4TransferLabel}
\alias{prepareDotPlot4TransferLabel}
\title{tfmat: transfer label score: ref by query}
\usage{
prepareDotPlot4TransferLabel(
  tfmat,
  refOrder = NULL,
  names = c("row", "column", "score"),
  ignoreEmptyRef = TRUE
)
}
\description{
tfmat: transfer label score: ref by query
}
