% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseurat.R
\name{read10X.seurat}
\alias{read10X.seurat}
\title{10X gene mtx to Seurat object}
\usage{
read10X.seurat(
  dir_of_10X,
  minCell = 3,
  minFeature = 100,
  project = "10X",
  mitoPattern = "^mt-"
)
}
\description{
10X gene mtx to Seurat object
}
