% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromHMM.R
\name{readChromHMMDenseBed}
\alias{readChromHMMDenseBed}
\title{Read dense bed file from ChromHMM.
Ignore first row since it's a title for showing in track.}
\usage{
readChromHMMDenseBed(densefnm)
}
\arguments{
\item{densefnm}{chromHMM segmentation dense fnm}
}
\value{
data.frame
}
\description{
Read dense bed file from ChromHMM.
Ignore first row since it's a title for showing in track.
}
