% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homer.R
\name{readHomerTSS}
\alias{readHomerTSS}
\title{Read Homer(i.e, Homer2) genome TSS file into data.frame.}
\usage{
readHomerTSS(homertssf, bufferSize = 2000)
}
\arguments{
\item{homertssf}{full path of homer tss file.
For example, if the genome is mm10, it should be under
installed homer directory, "data/genomes/mm10/mm10.tss"}
}
\value{
data.frame
}
\description{
\enumerate{
\item Homer uses 2000 as buffersize to extend the raw TSS, which we
will get the raw single base pair annotation.
\item Homer use 1 as "-" strand and 0 as "+".
\item The first column is RefSeq RNA id, so multiple isoforms shared
the same TSS will also be listed.
Check "bin/parseGTF.pl" under the installed homer directory.
Check https://genome.ucsc.edu/FAQ/FAQformat.html for bed format.
}
}
\details{
ChromHMM also provides the coords. Under instealled dir,
"COORDS/mm10/RefSeqTSS.mm10.bed.gz" as one example.
It only records the single-base, unique (no RefSeq id),
and NOT-ordered TSS. Both NR (non-coding) and NM (coding)
RNAs are recorded.
}
