% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_atac.R
\name{readSummitAnnot}
\alias{readSummitAnnot}
\title{Get chromHMM annotations for ATAC-seq peaks.}
\usage{
readSummitAnnot(scfnm, promoter, sc = NULL)
}
\arguments{
\item{scfnm}{state-mapped ATAC-summit file.
Under \link{projd}/06.ChromHMM/out/CREAnnot250429
the filename is like \link{subclass}.CRE.annotBySummit.bed}

\item{promoter}{GRobject  see \link{getHomerPromoterGR} as example}

\item{sc}{subclass name}
}
\value{
data.frame with columns CRE, state,
pd (short for proximal or distal)
}
\description{
Get chromHMM annotations for ATAC-seq peaks.
}
