% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{resizeIgnoreStrand}
\alias{resizeIgnoreStrand}
\title{Resize the GenomicRanges with ext_size and ignore the strand info.}
\usage{
resizeIgnoreStrand(gr, fix = "center", extsize = 2500)
}
\arguments{
\item{gr}{GenomicRanges}

\item{fix}{str, can be c("center", "start", "end", "twoside")}

\item{extsize}{int, default 2500}
}
\value{
GenomicRanges
}
\description{
Resize the GenomicRanges with ext_size and ignore the strand info.
}
