% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_atac.R
\name{retrieveCRE}
\alias{retrieveCRE}
\title{Filter deATAC based on log2 fold change}
\usage{
retrieveCRE(sc, deATAC, log2fd = 0.2, n = 99999999L)
}
\arguments{
\item{sc}{PairedTag-format subclass}

\item{deATAC}{\link{snATAChrADEfnm}
The CREs has already be filtered by q-value 0.05, though
we should do it here for raw DE files.}

\item{log2fd}{0.2 by default}

\item{n}{number of top ranked CREs}
}
\value{
vector of CREs (chr1:0-1 format)
CREs will be sorted based on log2fd
}
\description{
Filter deATAC based on log2 fold change
}
