% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Ref:https://stackoverflow.com/questions/31574480/rgb-to-hex-converter}
\usage{
rgb2hex(rgbmat)
}
\arguments{
\item{rgbmat:}{3 by n}
}
\description{
Ref:https://stackoverflow.com/questions/31574480/rgb-to-hex-converter
}
