% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveCPM2h5}
\alias{saveCPM2h5}
\title{Save a matrix into hdf5 format.}
\usage{
saveCPM2h5(mat, peaks, clusters, outf)
}
\arguments{
\item{mat}{matrix, group by feature}

\item{peaks}{colnames of the matrix}

\item{clusters}{rownames of the matrix}

\item{outf}{output filename}
}
\value{
NULL/Nothing
}
\description{
In the hdf5 file, a group X is created, which
has three fileds, mat, colnames, rownames
In Python h5py package, we can access the mat by
\verb{f['X']['mat'][()]}, where f is a hdf5 file handle.
NOTE: hdf5r will transpose the mat
https://github.com/hhoeflin/hdf5r/issues/81
}
