% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupLogging}
\alias{setupLogging}
\title{logging with standard out and error to log file
generated by ChatGPT4
if used in REPL, remember to call
sink(NULL, type = "output")
sink(NULL, type = "message")
stop redicting output to log file}
\usage{
setupLogging(log_file_path, threshold = logger::TRACE)
}
\description{
logging with standard out and error to log file
generated by ChatGPT4
if used in REPL, remember to call
sink(NULL, type = "output")
sink(NULL, type = "message")
stop redicting output to log file
}
