% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{setVariableFeatures}
\alias{setVariableFeatures}
\title{depends on isOnDiskMat.Seurat, calVarOfFea.Seuat}
\usage{
setVariableFeatures(
  ref,
  query,
  features = NULL,
  onlayer = "counts",
  nfeatures = 2000,
  on = c("ref", "all"),
  eps = 1e-04
)
}
\description{
depends on isOnDiskMat.Seurat, calVarOfFea.Seuat
}
