% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{strGRSeq}
\alias{strGRSeq}
\title{Transform GenomicRanges Object to chr:start=end format}
\usage{
strGRSeq(gr)
}
\arguments{
\item{gr}{GenomicRanges Object}
}
\value{
Array of characters
}
\description{
Transform GenomicRanges Object to chr:start=end format
}
