% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{transformGenomicRange2DataFrame}
\alias{transformGenomicRange2DataFrame}
\title{transformGenomicRange2DataFrame}
\usage{
transformGenomicRange2DataFrame(gr, zeroBased = TRUE, default_strand = ".")
}
\description{
transformGenomicRange2DataFrame
}
