% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{transformGR2SimpleBed}
\alias{transformGR2SimpleBed}
\title{Transform GenomicRanges Object to
three-column bed (simple bed)}
\usage{
transformGR2SimpleBed(gr)
}
\arguments{
\item{gr}{GenomicRanges object}
}
\value{
data.frame three columns
}
\description{
Transform GenomicRanges Object to
three-column bed (simple bed)
}
