% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRanges.R
\name{transformRegion2GR}
\alias{transformRegion2GR}
\title{tranformRegion2GR}
\usage{
transformRegion2GR(region)
}
\arguments{
\item{region}{vector of "chr1:1-2" format}
}
\value{
data.frame with the column names
c(chr, startFrom, endTo, name)
}
\description{
tranformRegion2GR
}
